Move all fs-extra usages to the abstraction
